organize Urls -> Models
