+++ src/FilterLists.Agent/Features/Urls/Models/EntityUrlDtos/LicenseUrls.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using JetBrains.Annotations;
 
-namespace FilterLists.Agent.Features.Urls.Models
+namespace FilterLists.Agent.Features.Urls.Models.EntityUrlDtos
 {
     [UsedImplicitly]
     public class LicenseUrls
