+++ src/FilterLists.Agent/Features/Urls/ValidateAllUrls.cs
@@ -3,7 +3,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using FilterLists.Agent.Core.Interfaces;
-using FilterLists.Agent.Features.Urls.Models;
+using FilterLists.Agent.Features.Urls.Models.EntityUrlDtos;
+using FilterLists.Agent.Features.Urls.Models.ValidationResults;
 using MediatR;
 
 namespace FilterLists.Agent.Features.Urls
@@ -30,27 +31,32 @@ namespace FilterLists.Agent.Features.Urls
                 var results = new List<DataFileUrlValidationResults>();
 
                 var licenseUrls = await _repo.GetAllAsync<LicenseUrls>();
-                var licenseUrlErrors = (await _mediator.Send(new ValidateUrls.Command(licenseUrls), cancellationToken)).ToList();
+                var licenseUrlErrors = (await _mediator.Send(new ValidateUrls.Command(licenseUrls), cancellationToken))
+                    .ToList();
                 if (licenseUrlErrors.Any())
                     results.Add(new DataFileUrlValidationResults("License.json", licenseUrlErrors));
 
                 var listUrls = await _repo.GetAllAsync<ListUrls>();
-                var listUrlErrors = (await _mediator.Send(new ValidateUrls.Command(listUrls), cancellationToken)).ToList();
+                var listUrlErrors =
+                    (await _mediator.Send(new ValidateUrls.Command(listUrls), cancellationToken)).ToList();
                 if (listUrlErrors.Any())
                     results.Add(new DataFileUrlValidationResults("FilterList.json", listUrlErrors));
 
                 var maintainerUrls = await _repo.GetAllAsync<MaintainerUrls>();
-                var maintainerUrlErrors = (await _mediator.Send(new ValidateUrls.Command(maintainerUrls), cancellationToken)).ToList();
+                var maintainerUrlErrors =
+                    (await _mediator.Send(new ValidateUrls.Command(maintainerUrls), cancellationToken)).ToList();
                 if (maintainerUrlErrors.Any())
                     results.Add(new DataFileUrlValidationResults("Maintainer.json", maintainerUrlErrors));
 
                 var softwareUrls = await _repo.GetAllAsync<SoftwareUrls>();
-                var softwareUrlErrors = (await _mediator.Send(new ValidateUrls.Command(softwareUrls), cancellationToken)).ToList();
+                var softwareUrlErrors =
+                    (await _mediator.Send(new ValidateUrls.Command(softwareUrls), cancellationToken)).ToList();
                 if (softwareUrlErrors.Any())
                     results.Add(new DataFileUrlValidationResults("Software.json", softwareUrlErrors));
 
                 var syntaxUrls = await _repo.GetAllAsync<SyntaxUrls>();
-                var syntaxUrlErrors = (await _mediator.Send(new ValidateUrls.Command(syntaxUrls), cancellationToken)).ToList();
+                var syntaxUrlErrors = (await _mediator.Send(new ValidateUrls.Command(syntaxUrls), cancellationToken))
+                    .ToList();
                 if (syntaxUrlErrors.Any())
                     results.Add(new DataFileUrlValidationResults("Syntax.json", syntaxUrlErrors));
