+++ src/FilterLists.Agent/Infrastructure/Repositories/UrlRepository.cs
@@ -4,7 +4,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Threading.Tasks;
 using FilterLists.Agent.Core.Interfaces;
-using FilterLists.Agent.Features.Urls.Models;
+using FilterLists.Agent.Features.Urls.Models.EntityUrlDtos;
 using FilterLists.Agent.Infrastructure.Clients;
 using RestSharp;
